Rename confluent_kafka_output to output
