+++ tests/unit/connector/test_confluent_kafka.py
@@ -21,7 +21,7 @@ from logprep.connector.confluent_kafka.input import (
     ConfluentKafkaInput,
     ConfluentKafkaInputFactory,
 )
-from logprep.connector.confluent_kafka.confluent_kafka_output import (
+from logprep.connector.confluent_kafka.output import (
     ConfluentKafkaOutput,
     UnknownOptionError,
     ConfluentKafkaOutputFactory,
@@ -862,7 +862,7 @@ class TestConfluentKafka:
         kafka.shut_down()
         assert kafka._consumer is None
 
-    @mock.patch("logprep.connector.confluent_kafka.confluent_kafka_output.Producer")
+    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
     def test_store_custom_calls_producer_flush_on_buffererror(self, mock_producer):
         config = deepcopy(TestConfluentKafkaFactory.valid_configuration)
         kafka = ConfluentKafkaOutputFactory.create_from_configuration(config)
@@ -873,7 +873,7 @@ class TestConfluentKafka:
         kafka.store_custom({"message": "does not matter"}, "doesnotcare")
         kafka._producer.flush.assert_called()
 
-    @mock.patch("logprep.connector.confluent_kafka.confluent_kafka_output.Producer")
+    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
     def test_store_failed_calls_producer_flush_on_buffererror(self, mock_producer):
         config = deepcopy(TestConfluentKafkaFactory.valid_configuration)
         kafka = ConfluentKafkaOutputFactory.create_from_configuration(config)
@@ -886,7 +886,7 @@ class TestConfluentKafka:
         )
         kafka._producer.flush.assert_called()
 
-    @mock.patch("logprep.connector.confluent_kafka.confluent_kafka_output.Producer")
+    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
     def test_shut_down_calls_producer_flush(self, mock_producer):
         config = deepcopy(TestConfluentKafkaFactory.valid_configuration)
         kafka = ConfluentKafkaOutputFactory.create_from_configuration(config)
@@ -894,7 +894,7 @@ class TestConfluentKafka:
         kafka.shut_down()
         mock_producer.flush.assert_called()
 
-    @mock.patch("logprep.connector.confluent_kafka.confluent_kafka_output.Producer")
+    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
     def test_shut_down_sets_producer_to_none(self, mock_producer):
         config = deepcopy(TestConfluentKafkaFactory.valid_configuration)
         kafka = ConfluentKafkaOutputFactory.create_from_configuration(config)
