+++ tests/unit/framework/test_pipeline.py
@@ -543,7 +543,7 @@ class TestPipeline(ConfigurationForTests):
         assert test_event == {"any": "content", "version_info": "something random"}
 
     @mock.patch("logprep.connector.confluent_kafka.input.Consumer")
-    @mock.patch("logprep.connector.confluent_kafka.confluent_kafka_output.Producer")
+    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
     def test_pipeline_kafka_batch_finished_callback_is_called(self, _, __, ___):
         logprep_config = {
             "version": 1,
@@ -593,7 +593,7 @@ class TestPipeline(ConfigurationForTests):
         pipeline._input.batch_finished_callback.assert_called()
 
     @mock.patch("logprep.connector.confluent_kafka.input.Consumer")
-    @mock.patch("logprep.connector.confluent_kafka.confluent_kafka_output.Producer")
+    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
     def test_pipeline_kafka_batch_finished_callback_calls_store_offsets(self, _, __, ___):
         logprep_config = {
             "version": 1,
@@ -644,7 +644,7 @@ class TestPipeline(ConfigurationForTests):
         pipeline._input._consumer.store_offsets.assert_called()
 
     @mock.patch("logprep.connector.confluent_kafka.input.Consumer")
-    @mock.patch("logprep.connector.confluent_kafka.confluent_kafka_output.Producer")
+    @mock.patch("logprep.connector.confluent_kafka.output.Producer")
     def test_pipeline_kafka_batch_finished_callback_calls_store_offsets_with_message(
         self, _, __, ___
     ):
